+++ .gitignore
@@ -11,7 +11,7 @@ target/
 *.iws
 *.iml
 *.ipr
-.idea
+
 ### Eclipse ###
 .apt_generated
 .classpath
