+++ src/main/java/urfu/Main.java
@@ -1,9 +1,11 @@
+package urfu;
+
 import org.telegram.telegrambots.longpolling.TelegramBotsLongPollingApplication;
 
 public class Main {
     public static void main(String[] args) {
         ListOfOrders listOfOrders = ListOfOrders.INSTANCE;
-        //UsersState usersState = UsersState.INSTANCE;
+        //urfu.UsersState usersState = urfu.UsersState.INSTANCE;
 
         String botToken = System.getenv("TG_TOKEN");
         try (TelegramBotsLongPollingApplication botsApplication = new TelegramBotsLongPollingApplication()) {
@@ -12,5 +14,6 @@ public class Main {
         } catch (Exception e) {
             e.printStackTrace();
         }
+
     }
 }
