--- src/main/resources/menu.json
@@ -1,5 +1,0 @@
-{
-  "ЛюляКебаб":170,
-  "Напиток":140,
-  "Шаурма":250
-}
